requirements-1.txt
amqp==5.2.0
aniso8601==9.0.1
async-timeout==4.0.3
Babel==2.13.1
bcrypt==4.1.1
billiard==4.2.0
bleach==6.1.0
blinker==1.7.0
cached-property==1.5.2
cachelib==0.9.0
celery==5.3.6
checksum==1.0.4
checksumdir==1.2.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
distlib==0.3.6
dnspython==2.4.2
email-validator==2.1.0.post1
escape==1.1
filelock==3.12.0
Flask==3.0.0
Flask-BabelEx==0.9.4
Flask-Bcrypt==1.0.1
Flask-Caching==2.1.0
Flask-Cors==4.0.0
Flask-JWT-Extended==4.6.0
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-Principal==0.4.0
Flask-RESTful==0.3.10
Flask-Security-Too==5.3.2
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
fonttools==4.51.0
greenlet==3.0.1
idna==3.6
importlib-resources==6.1.1
inflection==0.5.1
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.5
kombu==5.3.7
MarkupSafe==2.1.3
matplotlib==3.8.4
mypy-extensions==1.0.0
numpy==1.26.4
packaging==24.0
pandas==2.2.2
passlib==1.7.4
pillow==10.3.0
platformdirs==3.5.1
prompt-toolkit==3.0.43
PyJWT==2.8.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2023.3.post1
redis==5.0.3
seaborn==0.13.2
six==1.16.0
speaklater==1.3
SQLAlchemy==2.0.23
sqlalchemy-orm==1.2.10
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2024.1
vine==5.1.0
virtualenv==20.23.0
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.0.1
WTForms==3.1.1
